Style: 응답 관련 클래스 포맷 통일, 패키지 이동
